+++ packages/uicore/src/components/InputWithIdentifier/InputWithIdentifier.tsx
@@ -48,6 +48,10 @@ export interface InputWithIdentifierProps {
    * @default 63
    */
   maxInput?: number
+  /**
+   * @default true
+   */
+  useUnversialToolTipId?: boolean
 }
 
 // https://harness.atlassian.net/wiki/spaces/CDNG/pages/736200458/Entity+Identifier
@@ -69,14 +73,19 @@ export const InputWithIdentifier: React.FC<InputWithIdentifierProps> = props => 
     idName = 'identifier',
     inputGroupProps,
     isIdentifierEditable = true,
-    maxInput = 63
+    maxInput = 63,
+    useUnversialToolTipId = true
   } = props
   const [editable, setEditable] = useState(false)
   const [userModifiedIdentifier, setUserModifiedIdentifier] = useState(false)
   const identifier = get(formik.values, idName) as string
   const [currentEditField, setCurrentEditField] = useState(inputLabel)
   const tooltipContext = useContext(FormikTooltipContext)
-  const dataTooltipId = tooltipContext?.formName ? `${tooltipContext?.formName}_${idName}` : ''
+  const dataTooltipId = !useUnversialToolTipId
+    ? tooltipContext?.formName
+      ? `${tooltipContext?.formName}_${idName}`
+      : ''
+    : 'universalInputNameWithId'
 
   return (
     <div className={cx(css.txtNameContainer, textCss.main)}>