+++ packages/uicore/webpack.config.js
@@ -10,10 +10,12 @@
 //
 
 const path = require('path')
-const nodeExternals = require('webpack-node-externals')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin')
+const packageJSON = require('./package.json')
 const isDev = process.env.NODE_ENV === 'development'
+
+const externals = Object.keys(packageJSON.peerDependencies).reduce((p, c) => ({ ...p, [c]: `commonjs ${c}` }), {})
 
 module.exports = {
   mode: isDev ? 'development' : 'production',
@@ -86,7 +88,7 @@ module.exports = {
     libraryTarget: 'umd'
   },
 
-  externals: [nodeExternals()],
+  externals,
 
   plugins: [
     new MiniCssExtractPlugin({