+++ packages/uicore/CHANGELOG.md
@@ -3,6 +3,221 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.39.0](https://github.com/harness/uicore/compare/@harness/uicore@3.38.0...@harness/uicore@3.39.0) (2022-06-22)
+
+**Note:** Version bump only for package @harness/uicore
+
+
+
+
+
+# [3.38.0](https://github.com/harness/uicore/compare/@harness/uicore@3.37.0...@harness/uicore@3.38.0) (2022-06-22)
+
+**Note:** Version bump only for package @harness/uicore
+
+
+
+
+
+# [3.37.0](https://github.com/harness/uicore/compare/@harness/uicore@3.36.0...@harness/uicore@3.37.0) (2022-06-21)
+
+**Note:** Version bump only for package @harness/uicore
+
+
+
+
+
+# [3.36.0](https://github.com/harness/uicore/compare/@harness/uicore@3.35.0...@harness/uicore@3.36.0) (2022-06-21)
+
+**Note:** Version bump only for package @harness/uicore
+
+
+
+
+
+# [3.35.0](https://github.com/harness/uicore/compare/@harness/uicore@3.34.0...@harness/uicore@3.35.0) (2022-06-21)
+
+**Note:** Version bump only for package @harness/uicore
+
+
+
+
+
+# [3.34.0](https://github.com/harness/uicore/compare/@harness/uicore@3.33.0...@harness/uicore@3.34.0) (2022-06-20)
+
+**Note:** Version bump only for package @harness/uicore
+
+
+
+
+
+# [3.33.0](https://github.com/harness/uicore/compare/@harness/uicore@3.32.0...@harness/uicore@3.33.0) (2022-06-20)
+
+
+### Bug Fixes
+
+* SplitButton css specificity for secondary variation ([#719](https://github.com/harness/uicore/issues/719)) ([b9adb8a](https://github.com/harness/uicore/commit/b9adb8a2d2511ba651c4e59e54fd89726b53a3d3))
+
+
+
+
+
+# [3.32.0](https://github.com/harness/uicore/compare/@harness/uicore@3.31.0...@harness/uicore@3.32.0) (2022-06-20)
+
+
+### Features
+
+* [CDS-38779]: add clear button to select with sub menu ([#718](https://github.com/harness/uicore/issues/718)) ([ed51d3d](https://github.com/harness/uicore/commit/ed51d3d69351426bc3942ae8bf0d8c74339e91c2))
+
+
+
+
+
+# [3.31.0](https://github.com/harness/uicore/compare/@harness/uicore@3.30.0...@harness/uicore@3.31.0) (2022-06-16)
+
+
+### Features
+
+* add disabled state in accordion panel and css cleanup ([#717](https://github.com/harness/uicore/issues/717)) ([0c46b02](https://github.com/harness/uicore/commit/0c46b020caa467829c3f8c25d423234f5b95df8e))
+
+
+
+
+
+# [3.30.0](https://github.com/harness/uicore/compare/@harness/uicore@3.29.0...@harness/uicore@3.30.0) (2022-06-15)
+
+
+### Features
+
+* add onChange on Accordion ([#716](https://github.com/harness/uicore/issues/716)) ([f3486ba](https://github.com/harness/uicore/commit/f3486ba2e357356454c56359add69f59c3705287))
+
+
+
+
+
+# [3.29.0](https://github.com/harness/uicore/compare/@harness/uicore@3.28.0...@harness/uicore@3.29.0) (2022-06-15)
+
+
+### Bug Fixes
+
+* pass tooltipProps to second button only in SplitButton component ([#715](https://github.com/harness/uicore/issues/715)) ([d5a2ee0](https://github.com/harness/uicore/commit/d5a2ee0bec1a4494acaee4baea0ffafc612a40b0))
+
+
+
+
+
+# [3.28.0](https://github.com/harness/uicore/compare/@harness/uicore@3.27.0...@harness/uicore@3.28.0) (2022-06-15)
+
+
+### Bug Fixes
+
+* add display flex to SplitButton ([#714](https://github.com/harness/uicore/issues/714)) ([8c82fae](https://github.com/harness/uicore/commit/8c82faef36716ad416205141cc6a06004e398fa0))
+
+
+
+
+
+# [3.27.0](https://github.com/harness/uicore/compare/@harness/uicore@3.26.0...@harness/uicore@3.27.0) (2022-06-14)
+
+
+### Bug Fixes
+
+* value parameter added to selectWithSubmenu ([#713](https://github.com/harness/uicore/issues/713)) ([3226137](https://github.com/harness/uicore/commit/3226137f73bcd809098cf3797f3181d6d66fba3c))
+
+
+
+
+
+# [3.26.0](https://github.com/harness/uicore/compare/@harness/uicore@3.25.0...@harness/uicore@3.26.0) (2022-06-14)
+
+
+### Features
+
+* [CDS-38820]: provide props in PillToggle component so that consumers show disable warning icon and tooltip ([#712](https://github.com/harness/uicore/issues/712)) ([658d91d](https://github.com/harness/uicore/commit/658d91d995fb91b9fba1604a54eed252e9fddd82))
+
+
+
+
+
+# [3.25.0](https://github.com/harness/uicore/compare/@harness/uicore@3.24.0...@harness/uicore@3.25.0) (2022-06-13)
+
+**Note:** Version bump only for package @harness/uicore
+
+
+
+
+
+# [3.24.0](https://github.com/harness/uicore/compare/@harness/uicore@3.23.0...@harness/uicore@3.24.0) (2022-06-13)
+
+
+### Features
+
+* [CDS-38800]: split button disabled mode and style fixes ([#708](https://github.com/harness/uicore/issues/708)) ([08b6d54](https://github.com/harness/uicore/commit/08b6d5424bbdb354d0e0cc4eb520e959aaf0b619))
+
+
+
+
+
+# [3.23.0](https://github.com/harness/uicore/compare/@harness/uicore@3.22.0...@harness/uicore@3.23.0) (2022-06-09)
+
+**Note:** Version bump only for package @harness/uicore
+
+
+
+
+
+# [3.22.0](https://github.com/harness/uicore/compare/@harness/uicore@3.21.0...@harness/uicore@3.22.0) (2022-06-09)
+
+
+### Bug Fixes
+
+* formik bundling issue ([#706](https://github.com/harness/uicore/issues/706)) ([78e72b5](https://github.com/harness/uicore/commit/78e72b5ed06276f5964c893e05a0810fe40ff4cb))
+
+
+
+
+
+# [3.21.0](https://github.com/harness/uicore/compare/@harness/uicore@3.20.0...@harness/uicore@3.21.0) (2022-06-08)
+
+
+### Bug Fixes
+
+* [PIE-3783]: upgrades to SelectWithSubmenu ([#703](https://github.com/harness/uicore/issues/703)) ([cc0193d](https://github.com/harness/uicore/commit/cc0193d72499f08ec3b70ae01c24b480cdf1b08b))
+
+
+
+
+
+# [3.20.0](https://github.com/harness/uicore/compare/@harness/uicore@3.19.0...@harness/uicore@3.20.0) (2022-06-08)
+
+**Note:** Version bump only for package @harness/uicore
+
+
+
+
+
+# [3.19.0](https://github.com/harness/uicore/compare/@harness/uicore@3.18.0...@harness/uicore@3.19.0) (2022-06-08)
+
+
+### Features
+
+* [PL-25668]: Universal tooltipid for all id component ([#701](https://github.com/harness/uicore/issues/701)) ([77d83d6](https://github.com/harness/uicore/commit/77d83d6adeb248770bfb343d8a546db4cc43e5bd))
+
+
+
+
+
+# [3.18.0](https://github.com/harness/uicore/compare/@harness/uicore@3.17.0...@harness/uicore@3.18.0) (2022-06-08)
+
+
+### Bug Fixes
+
+* externals config in webpack ([#700](https://github.com/harness/uicore/issues/700)) ([ffc5353](https://github.com/harness/uicore/commit/ffc535379c8e46ccbc4aa0c847bf941350e01e20))
+
+
+
+
+
 # [3.17.0](https://github.com/harness/uicore/compare/@harness/uicore@3.16.0...@harness/uicore@3.17.0) (2022-06-07)
 
 **Note:** Version bump only for package @harness/uicore